+++ src/model/http.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize)]
+pub struct Response<T> {
+    pub correlation_id: String,
+    pub message: String,
+    pub error: String,
+    pub data: Option<T>,
+}
