+++ src/model/item.rs
@@ -1,0 +1,7 @@
+use serde::Serialize;
+
+#[derive(Serialize, Clone)]
+pub struct Item {
+    pub id: String,
+    pub name: String,
+}
